♻️ refactor: 유틸함수 생성 및 적용 (Jira Ticket : CCC-41)
