+++ components/modal-template/EmailCopyModal.tsx
@@ -7,6 +7,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
+import copyText from '@/utils/copyText';
 // import { Member } from '@ccc-types';
 import Image from 'next/image';
 import React, { ReactNode } from 'react';
@@ -23,17 +24,6 @@ const userMockData = {
 };
 
 function EmailCopyModal({ children }: { children: ReactNode }) {
-  const copyText = (text: string) => {
-    navigator.clipboard
-      .writeText(text)
-      .then(() => {
-        alert('이메일이 복사되었습니다!');
-      })
-      .catch(() => {
-        alert('이메일 복사에 실패하였습니다!');
-      });
-  };
-
   return (
     <Dialog>
       <DialogTrigger asChild>{children}</DialogTrigger>
